Fix replication error on multi transaction
There was an error upon replication of multixact if it is a first on the page.

Previously, pd_multi_base was not wrote in WAL and was set to 0 upon WAL reply
on replica. This caused replica to panic. Fix this by adding pd_multi_base of a
page into WAL.

Appropriate tap test is added.