Use 64-bit multixact offsets
Switching to 64-bit multitransaction offsets removes wraparound and the
2^32 limit on their total number.

On the other hand, this move is inevitable in increased disc space
utilisation.  Fortunately, multitransaction offsets rise monotonically
and without gaps.  To conserve disc space consumed by segments, we
write a 64-bit "base" at the start of each page, which also serves as
the page's first offset.  All subsequent offsets on the page are
calculated relative to this "base".

Author: Maxim Orlov <orlovmg@gmail.com>